refactor: change api
